Matt dev

+++ test/test_get_xml.py
@@ -11,11 +11,15 @@ def clear_old_test_data():
     folder = '{}{}'.format(os.getcwd(), '/data_TEST/')
     for the_file in os.listdir(folder):
         file_path = os.path.join(folder, the_file)
+        leave_files = ('1_transform_test.xml',
+                       '10_transform_test.xml',
+                       '20_transform_test.xml',)
         try:
-            if os.path.isfile(file_path):
+            if os.path.isfile(file_path) and the_file not in \
+                    leave_files:
                 os.unlink(file_path)
         except Exception as e:
-            print(e)
+            print e
 
 
 class TestGetXML(unittest.TestCase):
@@ -73,7 +77,3 @@ class TestGetXML(unittest.TestCase):
         response code is returned and the text value is None
         """
         self.assertRaises(Exception, self.badGX.download_xml())
-
-
-suite = unittest.TestLoader().loadTestsFromTestCase(TestGetXML)
-unittest.TextTestRunner(verbosity=2).run(suite)
